Rename TranslationUpdateCommand to TranslationExtract command to match the command name
